+++ sigar_windows.go
@@ -136,11 +136,16 @@ func (self *FileSystemList) Get() error {
 		if err != nil {
 			return errors.Wrapf(err, "GetDriveType failed")
 		}
+		fsType, err := windows.GetFilesystemType(drive)
+		if err != nil {
+			return errors.Wrapf(err, "GetFilesystemType failed")
+		}
 
 		self.List = append(self.List, FileSystem{
-			DirName:  drive,
-			DevName:  drive,
-			TypeName: dt.String(),
+			DirName:     drive,
+			DevName:     drive,
+			TypeName:    dt.String(),
+			SysTypeName: fsType,
 		})
 	}
 	return nil